requirements.txt
# Discord and config handling
discord.py
toml

# Storage handling
asyncpg
aioredis
